Refactor Music Player: Extract PlayMusic Component and Implement Favorite Toggle
